assignment2/web1/requirements.txt
flask
gunicorn
requests
flask-mysql
mysql-connector-python
